+++ sdk/dotnet/Inputs/ContainerNetworkDataGetArgs.cs
@@ -19,6 +19,12 @@ namespace Pulumi.Docker.Inputs
         [Input("gateway")]
         public Input<string>? Gateway { get; set; }
 
+        [Input("globalIpv6Address")]
+        public Input<string>? GlobalIpv6Address { get; set; }
+
+        [Input("globalIpv6PrefixLength")]
+        public Input<int>? GlobalIpv6PrefixLength { get; set; }
+
         /// <summary>
         /// *Deprecated:* Use `network_data` instead. The IP address of the container's first network it.
         /// </summary>
@@ -31,6 +37,9 @@ namespace Pulumi.Docker.Inputs
         /// </summary>
         [Input("ipPrefixLength")]
         public Input<int>? IpPrefixLength { get; set; }
+
+        [Input("ipv6Gateway")]
+        public Input<string>? Ipv6Gateway { get; set; }
 
         [Input("networkName")]
         public Input<string>? NetworkName { get; set; }