+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (Unreleased)
-_(none)_
+
+* Upgrade to v2.8.0 of the Docker Terraform Provider
+
 
 ---